Added seminar_2 tasks and solution with diff

+++ Examples/Seminar_2/Program.cs
@@ -135,9 +135,9 @@ class Program
 */
 
 Console.WriteLine("Input first number");
-int num1 = Convert.ToInt32(Console.ReadLine());
+double num1 = Convert.ToInt32(Console.ReadLine());
 Console.WriteLine("Input second number");
-int num2 = Convert.ToInt32(Console.ReadLine());
+double num2 = Convert.ToInt32(Console.ReadLine());
 
 if(num1 == num2*num2 || num2 == num1*num1)
 {
